[ErrorHandling] Exit if processing fails.
If a message isn't ACK'd, then the RabbitMQ server will wait until the connection is closed before sending the message to another consumer.  What this means is that if we don't ACK then we have to, at the least, close the connection.

The exit status of 1 should ensure that Heroku thinks the app crashed, which will show up rather visibly in Logentries.  We can push messages into Hipchat if we desire.